requirements.txt
# Core ML
torch>=2.0
transformers>=4.36
datasets>=2.16
accelerate>=0.25
sentencepiece>=0.1.99
protobuf>=3.20

# Server
fastapi>=0.109
uvicorn[standard]>=0.27
pydantic>=2.0

# Data processing
scikit-learn>=1.4
pandas>=2.1
pyarrow>=14.0
datasketch>=1.6
numpy>=1.24

# Utilities
tqdm>=4.66
pyyaml>=6.0
requests>=2.31

# Testing
httpx>=0.26
pytest>=8.0
pytest-asyncio>=0.23

# Stage 3 - Claude API (optional)
anthropic>=0.40

# Evaluation & reporting
matplotlib>=3.8
seaborn>=0.13
